test(ethers-providers): move platform-independent tests to commonTest
Splits the suite by what each test actually needs, rather than leaving all of
it on the JVM.

Moved wholesale: RpcErrorTest and CallManyTypesTest, which only exercise
serialization. CallManyTypesTest loses its `@Language("JSON")` hints, since
those come from org.jetbrains:annotations and are JVM-only; the JSON literals
are unchanged.

Split: the combinator contexts of RpcRequestTests only use the suspending API,
so they become RpcCombinatorTest in commonTest, leaving the blocking and
CompletableFuture adapter tests behind. ProviderBuilderTest moves except for
its two `buildAwait` assertions, which become ProviderBuilderAwaitTest.

What stays, and why:
  - MockServer, PendingTransactionTest: okhttp3 MockWebServer is JVM-only
  - HttpClientTest, WsClientTest, ClientOwnershipTest: the ktor CIO engine is
    not available on every target
  - JsonRpcTestFactory: CompletableFuture assertions
  - RpcRequestTests, ProviderBuilderAwaitTest: the blocking API itself

commonTest is clean of java.*, okhttp3, CIO, runBlocking, CompletableFuture and
the blocking API. Migrating the rest needs ktor-client-mock, which is already
in the version catalog but unused, to replace MockWebServer.
