+++ ethers-providers/src/commonTest/kotlin/io/ethers/providers/RpcErrorTest.kt
